Fix 64 bit memory block tag identifier format; fix int-size_t issue from openssl

+++ src/tls_interface.cc
@@ -285,7 +285,7 @@ static EVP_PKEY *genKey( int kBits ) {
 		return NULL;
 	{
 		OSSL_PARAM params[2] = {
-			OSSL_PARAM_BN( OSSL_PKEY_PARAM_RSA_E, bne, BN_num_bytes( bne ) ),
+			OSSL_PARAM_BN( OSSL_PKEY_PARAM_RSA_E, bne, (size_t)BN_num_bytes( bne ) ),
 			OSSL_PARAM_END
 		};
 		if( EVP_PKEY_fromdata_init( ctx ) != 1 ||
